Fix deindex glossary

+++ lib/logaling/glossary_db.rb
@@ -63,8 +63,10 @@ module Logaling
 
     def deindex_glossary(glossary, glossary_source)
       delete_translations_by_glossary_source(glossary_source.source_path)
-      delete_glossary(glossary.name)
       delete_glossary_source(glossary_source.source_path)
+      unless glossary_source_of_the_same_project_exist?(glossary_source)
+        delete_glossary(glossary.name)
+      end
     end
 
     def deindex_glossary_source(glossary_source)
@@ -436,6 +438,25 @@ module Logaling
     def offline_index?
       # use online index if LOGALING_OFFLINE_INDEX=no
       ENV["LOGALING_OFFLINE_INDEX"] != "no"
+    end
+
+    def glossary_source_of_the_same_project_exist?(glossary_source)
+      glossary_source_num = 0
+      get_all_glossary_sources.each do |glossary_source_taken|
+        if glossary_source.belongs_to_personal? && glossary_source_taken.belongs_to_personal?
+          if glossary_source.glossary_name == glossary_source_taken.glossary_name
+            glossary_source_num = 1
+            break
+          end
+        elsif glossary_source.belongs_to_project? && glossary_source_taken.belongs_to_project?
+          if glossary_source.project_name == glossary_source_taken.project_name &&
+             glossary_source.glossary_name == glossary_source_taken.glossary_name
+            glossary_source_num = 1
+            break
+          end
+        end
+      end
+      glossary_source_num > 0 ? true : false
     end
   end
 end
